pyproject.toml
[project]
name = "airflow-dags"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Vijay Kiran", email = "mail@vijaykiran.com" }
]
dependencies = [
    "apache-airflow[celery]==2.10.0",
    "apache-airflow-providers-cncf-kubernetes>=8.4.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/airflow_dags"]
